Project files restructuring
